backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
fastapi==0.115.8
frozenlist==1.8.0
h11==0.16.0
idna==3.11
motor==3.7.0
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dotenv==1.0.1
python-multipart==0.0.22
qrcode==8.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.15.0
uvicorn==0.34.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
